Simplify and tidy code
